requirements.txt
requests
beautifulsoup4
lxml
jinja2
python-slugify
tzdata
